WIP Credit Attribution
Currently has placeholder credits in place for the two types of raster overlays. Currently uses dynamically aggregated strings list which is inefficient. Once a POC test can be done on the UI side, we should think about using a pre-loaded master list of credits and index into it on the fly.