Initialize the members with the proper defaults.
The tests assumed that the enum members are initialized
with "the first" enum value, if no explicit default value
was given. In the schema, the "type" is only specified in
every element of "anyOf" when the type is "integer" (but
not when it is "string"). Not sure who to blame here...