Add empty source file to Cesium3DTiles library.
Otherwise there's no .lib on a clean build, which breaks the build of
the tests. We could fix that by declaring the library an INTERFACE
library, but that would require further changes to our cmake config. No
source files in Cesium3DTiles is probably only temporary, so we'll just
work around it for now.