Don't require that parent overlay fully covers child to use it.
The immediate problem is that the implementation would falsely flag some
parent tiles as not fully covering children, and so we'd end up
rendering untextured grey areas on the globe.

But the bigger reason is because even a parent texture that doesn't fully
cover the child is probably a better to use than the ancestor tile (or
nothing at all). The texture will just be stretched to the edges.