Don't use static CRT on Windows.
Because that would require doing the same for vcpkg dependencies, which
is tricky because vcpkg doesn't provide an out-of-the-box triplet that
uses the static CRT in debug buids. We could create one, but it's not
worth the trouble. The static CRT was added to work around a GitHub
runner bug that has since been fixed.