Fix an unknown bug in rayOBBParametric by rewriting it.
There were tests, and the old implementation passed them. I added new
tests, and it passed those, too. But some real-world height queries were
returning way too many potential tiles, and I narrowed it down to a
problem with this intersection test. When I rewrote it "the long way",
the too-many-tiles problem disappeared, and all the tests still pass.

So I'm not quite sure what was wrong with the old implementation, but
I'm committing this working one before I write more tests and optimize
it.