Don't create tiles past level 30, because it won't work.
EllipsoidTilesetLoader can conceptually create tiles of any level. But
because we use a uint32_t for the tile X and Y coordinates, and because
tile level 31 has 2*2^31=2^32 tiles, we end up overflowing the uint32_t
and producing nonsense tile coordinates. So this change prevents the
loader from creating tiles past level 30.