Eliminate block of useless indices from EllipsoidTilesetLoader.
EllipsoidTilesetLoader was passing a size to the constructor for the
std::vector used to store the indices, and then it was also adding the
indices by calling `insert`. That resulted in a giant block of useless
(0,0,0) indices, followed by the useful ones created by the insert.

This change uses `reserve` instead of passing a size to the constructor.