Calculate I3dm instance rotations by converting basis to a matrix
The previous approach tried to create rotations directly from the up
and right vectors that encode the instance rotation, but it naïvely
didn't handle rotations of 180 degrees of the up and right vector. It
is reliable to create the rotation matrix, filling in the Z column
by doing a cross-product, and then let the algorithm for converting a
rotation matrix to a quaternion do its robust thing.