Register error for empty I3dm feature table
This was brought up in #1127 and not really addressed by other changes
to I3dm parsing. In the specific case of an i3dm file with completely
missing feature of feature binary sections i.e., a malformed file, no
error would be produced from the conversion process. This would result
in a down-stream crash.