Fix test failure on ARM64 macOS.
The test was moving the camera in such a way that that the tile was no longer in the view frustum. The only reason it was passing on any platform was because the camera would still be just inside the tile, and so the rule that we don't cull tiles under the camera would apply. But on ARM64 macOS, the camera ends up being very very slightly outside the tile instead of very slightly inside, and so the test fails.

The change here is to keep the tile inside the frustum, as was intended in the first place.