Re-add blend2d overlay port.
This is needed until https://github.com/microsoft/vcpkg/pull/47969 is
merged and we've updated to a vcpkg version that includes it.

Without it, builds with blend2d's `jit` feature disabled (Emscripten)
fail because the package erroneously attempts to
`find_dependency(asmjit)` even when there's no asmjit to find (because
it's disabled).