Suppress incorrect clang-tidy warning.
As of glm v1.0.2, and at least on Windows, an include of
`<glm/gtc/quaternion.hpp>` is required in order for
`glm::dmat4x4(rotationQuat)` to compile successfully. Otherwise, the
compiler complains about missing `mat4_cast`.

However, clang-tidy does not recognize that this header is required, and
throws a `misc-include-cleaner` warning. Possibly because it's not
required under clang.